+++ src/components/NewGameMenu/index.tsx
@@ -9,12 +9,12 @@ type props = {
 
 const sizes = [
   { radius: 1, label: "Small", description: "7 cells" },
-  { radius: 2, label: "Medium", description: "19 cells" },
-  { radius: 3, label: "Large", description: "37 cells" },
-  { radius: 4, label: "XL", description: "61 cells" },
+  { radius: 2, label: "Normal", description: "19 cells" },
+  // { radius: 3, label: "Large", description: "37 cells" },
+  // { radius: 4, label: "XL", description: "61 cells" },
 ];
 
-const NewGameMenu = ({ resetGameHandler, currentRadius = 1 }: props) => {
+const NewGameMenu = ({ resetGameHandler, currentRadius = 2 }: props) => {
   const [selected, setSelected] = useState(currentRadius);
   const cardRefs = useRef<(HTMLButtonElement | null)[]>([]);